Use std::size_t for accumulating hashes

+++ components/files/hash.cpp
@@ -11,7 +11,7 @@ namespace Files
 {
     std::uint64_t getHash(const std::string& fileName, std::istream& stream)
     {
-        std::uint64_t hash = std::hash<std::string> {}(fileName);
+        std::size_t hash = std::hash<std::string> {}(fileName);
         try
         {
             const auto start = stream.tellg();
